+++ cpp/src/arrow/testing/gtest_util.cc
@@ -176,8 +176,8 @@ void AssertZeroPadded(const Array& array) {
     if (buffer) {
       const int64_t padding = buffer->capacity() - buffer->size();
       if (padding > 0) {
-        std::vector<uint8_t> zeros(padding);
-        ASSERT_EQ(0, memcmp(buffer->data() + buffer->size(), zeros.data(), padding));
+        std::vector<uint8_t> zeros(static_cast<size_t>(padding));
+        ASSERT_EQ(0, memcmp(buffer->data() + buffer->size(), zeros.data(), static_cast<size_t>(padding)));
       }
     }
   }
